moved things from ServerImpostorViewModel to ImpostorViewModel
